Added monet readme
